select is working

+++ sqlite/src/page.rs
@@ -2,7 +2,7 @@
 #[derive(Debug)]
 pub struct Page {
     pub page_header: PageHeader,
-    pub cell_array: Vec<u64>,
+    pub cell_array: Vec<i64>,
 }
 
 #[derive(Debug)]
@@ -12,7 +12,7 @@ pub struct PageHeader {
     pub cell_count: u16,
     pub cell_content_area_offset: u16,
     pub fragmented_free_bytes: u8,
-    pub right_pointer: Option<u32>
+    pub right_pointer: Option<i64>
 }
 
 #[derive(Debug)]
